Decouple message queue, repo, and actor functionality from db instance w/ services
