+++ packages/pds/src/api/app/bsky/actor/getProfile.ts
@@ -4,7 +4,6 @@ import { countAll, actorWhereClause } from '../../../../db/util'
 import * as locals from '../../../../locals'
 import { getDeclarationSimple } from '../util'
 import ServerAuth from '../../../../auth'
-import { CID } from 'multiformats/cid'
 
 export default function (server: Server) {
   server.app.bsky.actor.getProfile({
