+++ packages/pds/src/stream/consumers/remove-upvote.ts
@@ -1,12 +1,14 @@
 import { sql } from 'kysely'
 import Database from '../../db'
+import { ActorService } from '../../services/actor'
 import { RemoveUpvote } from '../messages'
 import { Consumer } from '../types'
 
 export default class extends Consumer<RemoveUpvote> {
   async dispatch(ctx: { db: Database; message: RemoveUpvote }) {
     const { db, message } = ctx
-    const userScenes = await db.getScenesForUser(message.user)
+    const actorService = new ActorService(db)
+    const userScenes = await actorService.getScenesForUser(message.user)
     if (userScenes.length === 0) return
     await db.db
       .updateTable('scene_votes_on_post')
