+++ packages/pds/src/stream/types.ts
@@ -6,8 +6,9 @@ export type MessageOfType<T extends string = string> = {
 }
 
 export type Listener<M extends MessageOfType = MessageOfType> = (ctx: {
-  db: Database
   message: M
+  db: Database
+  messageQueue: MessageQueue
 }) => Promise<void | MessageOfType[]>
 
 export type Listenable<M extends MessageOfType = MessageOfType> =
