+++ packages/pds/tests/views/author-feed.test.ts
@@ -21,7 +21,7 @@ describe('pds author feed views', () => {
     close = server.close
     client = AtpApi.service(server.url)
     sc = new SeedClient(client)
-    await basicSeed(sc, server.db.messageQueue)
+    await basicSeed(sc, server.messageQueue)
     alice = sc.dids.alice
     bob = sc.dids.bob
     carol = sc.dids.carol
